refactor depth integration, support different integration functions (you can pass in different lambdas now), unify depth map and point cloud integrator, add unit tests
