+++ algorithms/dense-reconstruction/depth_integration/CMakeLists.txt
@@ -5,8 +5,7 @@ find_package(catkin_simple REQUIRED)
 catkin_simple(ALL_DEPS_REQUIRED)
 
 cs_add_library(${PROJECT_NAME}
-        src/depth-integration.cc
-        src/depth-map-integration.cc)
+        src/depth-integration.cc)
 
 SET(PROJECT_TEST_DATA "map_resources_test_data")
 add_custom_target(${PROJECT_TEST_DATA})
@@ -15,8 +14,12 @@ COMMAND rm -rf "./${PROJECT_TEST_DATA}/*" && tar -xvzf ${MAPLAB_TEST_DATA_DIR}/$
 
 catkin_add_gtest(test_voxblox_depth_integration test/test-voxblox-depth-integration.cc)
 target_link_libraries(test_voxblox_depth_integration ${PROJECT_NAME})
-maplab_import_test_maps(test_voxblox_depth_integration)
 add_dependencies(test_voxblox_depth_integration ${PROJECT_TEST_DATA})
+
+catkin_add_gtest(test_point_cloud_integration test/test-point-cloud-integration.cc)
+target_link_libraries(test_point_cloud_integration ${PROJECT_NAME})
+maplab_import_test_maps(test_point_cloud_integration)
+add_dependencies(test_point_cloud_integration ${PROJECT_TEST_DATA})
 
 cs_install()
 cs_export()
